third_party/blink/web_tests/external/wpt/css/css-cascade/scope-specificity-expected.txt

This is a testharness.js-based test.
[FAIL] @scope { & .b {  } } and :scope .b
  assert_equals: scoped + unscoped expected "1" but got "2"
Harness: the test ran to completion.
